Complete dashboard route decomposition
