--- apps/web/src/components/app/agents-workspace.tsx
@@ -1,17 +1,0 @@
-import { useEffect } from "react";
-
-export function AgentsWorkspace({
-  children,
-  onUnmount,
-}: {
-  children: React.ReactNode;
-  onUnmount: () => void;
-}): JSX.Element {
-  useEffect(() => {
-    return () => {
-      onUnmount();
-    };
-  }, [onUnmount]);
-
-  return <>{children}</>;
-}
