+++ apps/web/src/lib/agent-routes.ts
@@ -1,0 +1,14 @@
+export function agentRoute(agentId: string): string {
+  return `/agents/${agentId}`;
+}
+
+export function agentFeedbackRoute(agentId: string, itemId: number): string {
+  return `/agents/${agentId}/feedback/${itemId}`;
+}
+
+export function agentReviewRoute(
+  agentId: string,
+  reviewAgentId: string
+): string {
+  return `/agents/${agentId}/review/${reviewAgentId}`;
+}
